# talent-os/backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeadSubmit": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "interest_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interest Type"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name",
          "email",
          "message"
        ],
        "title": "LeadSubmit",
        "type": "object"
      },
...
      "QuizAnswerItem": {
        "properties": {
          "answer_index": {
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Answer Index",
            "type": "integer"
          },
          "question_id": {
            "title": "Question Id",
            "type": "integer"
          }
        },
        "required": [
          "question_id",
          "answer_index"
        ],
        "title": "QuizAnswerItem",
        "type": "object"
      },
      "QuizSubmitRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuizAnswerItem"
            },
            "minItems": 1,
            "title": "Answers",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "required": [
          "answers"
        ],
        "title": "QuizSubmitRequest",
        "type": "object"
      },
...
      "SalaryBenchmarkResponse": {
        "properties": {
          "currency": {
            "default": "EUR",
            "title": "Currency",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "p25": {
            "title": "P25",
            "type": "integer"
          },
          "p50": {
            "title": "P50",
            "type": "integer"
          },
          "p75": {
            "title": "P75",
            "type": "integer"
          },
          "p90": {
            "title": "P90",
            "type": "integer"
          },
          "role_title": {
            "title": "Role Title",
            "type": "string"
          },
          "sample_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Size"
          },
          "seniority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority"
          }
        },
        "required": [
          "role_title",
          "p25",
          "p50",
          "p75",
          "p90"
        ],
        "title": "SalaryBenchmarkResponse",
        "type": "object"
      },
...
      "SiteContentResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "section": {
            "title": "Section",
            "type": "string"
          }
        },
        "required": [
          "section",
          "items"
        ],
        "title": "SiteContentResponse",
        "type": "object"
      },
...
      "TalentpoolConfirmRequest": {
        "description": "Public: POST /api/public/talentpool-confirm. token is the raw,\nsingle-use value e-mailed to the candidate (only its sha256 hash is\never stored -- core.security.hash_token, same as WS-E.2's verify-email\nflow).",
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "TalentpoolConfirmRequest",
        "type": "object"
      },
...
      "TalentpoolOptinRequest": {
        "description": "Public: POST /api/public/talentpool-optin -- e-mail + consent tick\nfrom website/kandidaten.html or website/blog/post.html's CTA. Does not\nitself set anything on `candidates`; only issues a confirmation e-mail\n(routers/public.py talentpool_public_router). Consent only becomes\neffective once the token is confirmed via talentpool-confirm.",
        "properties": {
          "consent": {
            "title": "Consent",
            "type": "boolean"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "email",
          "consent",
          "scope",
          "source"
        ],
        "title": "TalentpoolOptinRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Multi-agent recruitment platform. PostgreSQL + asyncpg + FastAPI BackgroundTasks + Next.js.",
    "title": "Talent OS \u2014 Hermes Recruitment Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/public/talentpool-confirm": {
      "post": {
        "description": "Step 2: consumes the one-time token from the confirmation e-mail.\nOnly here does consent become effective on the candidates row --\ncreating it (source='talentpool_optin', no source_url \u2014 SOP \u00a71.5) if\nthis e-mail has no existing candidates row, or updating it in place\notherwise. lawful_basis is set to 'opt_in_talentpool' unless the\ncandidate already carries a different lawful_basis on file (never\nsilently overwritten \u2014 same rule as the portal endpoint).",
        "operationId": "talentpool_confirm_api_public_talentpool_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TalentpoolConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Talentpool Confirm",
        "tags": [
          "public-talentpool"
        ]
      }
    },
    "/api/public/talentpool-optin": {
      "post": {
        "description": "Step 1: e-mail + consent tick from website/kandidaten.html or the\nblog CTA. Never touches `candidates` directly -- only issues a\nconfirmation e-mail with a one-time token. Always returns the same\ngeneric message regardless of consent value or whether the e-mail is\nalready known, so this endpoint can't be used to enumerate e-mails or\nprobe existing candidates (same no-enumeration pattern as\nrouters/auth.py resend_verification).",
        "operationId": "talentpool_optin_api_public_talentpool_optin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TalentpoolOptinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Talentpool Optin",
        "tags": [
          "public-talentpool"
        ]
      }
    },
...
    "/api/v1/public/blog/{slug}": {
      "get": {
        "description": "Get a single published blog post by slug, including full bodies.",
        "operationId": "get_blog_post_api_v1_public_blog__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Blog Post",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/lead": {
      "post": {
        "description": "Submit a lead or contact form entry.",
        "operationId": "submit_lead_api_v1_public_lead_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Lead",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/quiz": {
      "get": {
        "description": "Get 12 random active quiz questions, balanced 3 per domain (mixed\ndifficulty). Never includes correct_index or explanations \u2014 those are\nonly returned by /quiz/submit, after grading.",
        "operationId": "get_quiz_api_v1_public_quiz_get",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "nl",
              "pattern": "^(nl|en)$",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quiz",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/quiz/submit": {
      "post": {
        "description": "Grade a quiz submission server-side and store it. Correct answers and\nexplanations are only ever exposed here, after submission \u2014 never by\nGET /quiz.",
        "operationId": "submit_quiz_api_v1_public_quiz_submit_post",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "nl",
              "pattern": "^(nl|en)$",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Quiz",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/salary-data": {
      "get": {
        "description": "Get salary benchmark data (public).",
        "operationId": "get_public_salary_data_api_v1_public_salary_data_get",
        "parameters": [
          {
            "in": "query",
            "name": "role_title",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Title"
            }
          },
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "seniority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SalaryBenchmarkResponse"
                  },
                  "title": "Response Get Public Salary Data Api V1 Public Salary Data Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Salary Data",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/site-content": {
      "get": {
        "description": "Get site content by section.",
        "operationId": "get_site_content_api_v1_public_site_content_get",
        "parameters": [
          {
            "description": "Content section (e.g. hero, features, about)",
            "in": "query",
            "name": "section",
            "required": true,
            "schema": {
              "description": "Content section (e.g. hero, features, about)",
              "title": "Section",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Site Content",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}